Refactored repairs to share more code
